+++ src/service/mod.rs
@@ -2,5 +2,6 @@ pub mod evaluate;
 pub mod factory;
 pub mod impl_http;
 
+pub mod destinations;
 pub mod origin_router;
 pub mod record;
